# app/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Startup State Atlas API
  version: "1.0"
  summary: Utah's agent-readable map of startup companies and resources.
  description: |
    The Startup State Atlas API powers the Founder Navigator, the Utah
    Startup Map, and the agent-native CLI/MCP surfaces.

    ## Authentication

    Two authentication modes are supported:

    - **Better Auth session cookie** (`auth.session_token`) — set by
      `/sign-in` and `/sign-up`. Used by humans clicking through the
      web UI. Roles: `owner`, `goeo_admin`, `superadmin`.
    - **`X-Atlas-Admin-Token`** header — service-account-style token
      validated against `env.ATLAS_ADMIN_TOKEN`. Used by the local
      CLI, the local stdio MCP server, and any external machine
      client. Bypasses Better Auth and assumes a privileged caller.

    Read endpoints are open to both. Write endpoints accept either.
    For owner-edit on `PATCH /api/v1/companies/{slug}`, the session
    user id must match `companies.claimed_by_user_id`.

    ## Error envelope

    Every non-2xx response uses the same shape:

    ```json
    { "error": { "code": "string", "message": "string", "details": "unknown?" } }
    ```

    Codes used: `bad_request`, `not_found`, `forbidden`, `unauthorized`,
    `conflict`, `internal`.

    ## Wire format

    All wire fields are `snake_case`. Internal TS uses `camelCase` and
    converts at the boundary via `lib/api-codec.ts`.
servers:
  - url: https://startup.utah.gov
...
security:
  - cookieAuth: []
  - atlasAdminToken: []
  - {}
...
paths:
...
  /api/v1/investor-profiles/{slug}:
    parameters:
      - in: path
        name: slug
        required: true
        schema: { type: string }
    get:
      tags: [investors]
      operationId: getInvestorProfileBySlug
      summary: Return the public investor card for a given slug.
      description: |
        Public — no auth required. Email is intentionally redacted; the
        only way to surface contact info is via an admin-accepted intro
        request.
      security: []
      responses:
        "200":
          description: Investor card.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvestorPublicCard" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [investors]
      operationId: patchInvestorProfile
      summary: Update an investor's public profile.
      description: |
        Three auth modes:
        - **Owner** (Better Auth session, `investor_profiles.user_id ===
          session.user.id`): can write the public-fields whitelist
          (display_name, bio, tagline, website, linkedin, slug while
          unverified). Editing display_name or website on a verified
          profile auto-flips `verification_status` back to `unverified`
          for re-review.
        - **Admin** (`goeo_admin` / `superadmin`): can also write
          `verification_status`. Admin edits do NOT auto-unverify.
        - **Machine** (`X-Atlas-Admin-Token`): same as admin.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvestorPublicPatch" }
      responses:
        "200":
          description: Updated investor card.
          content:
            application/json:
              schema:
                type: object
                required: [profile]
                properties:
                  profile: { $ref: "#/components/schemas/InvestorPublicCard" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Slug already taken.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/saved-companies:
    get:
      tags: [saved]
      operationId: listSavedCompanies
      summary: The caller's saved-companies watchlist.
      description: |
        Session required. Returns the calling investor's watchlist
        joined to companies. Returns an empty list if the caller has
        no investor profile.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Saved companies.
          content:
            application/json:
              schema:
                type: object
                required: [saved]
                properties:
                  saved:
                    type: array
                    items: { $ref: "#/components/schemas/SavedCompany" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [saved]
      operationId: saveCompany
      summary: Save a company to the caller's watchlist.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_id]
              properties:
                company_id:
                  type: string
                  pattern: "^co_[a-z0-9]+$"
                note: { type: string, nullable: true, maxLength: 1000 }
      responses:
        "201":
          description: Saved.
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: { type: string, pattern: "^sc_[a-z0-9]+$" }
                  status: { type: string, enum: [created] }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Already saved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    patch:
      tags: [saved]
      operationId: updateSavedCompanyNote
      summary: Update the note on a saved company.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_id, note]
              properties:
                company_id: { type: string, pattern: "^co_[a-z0-9]+$" }
                note: { type: string, nullable: true, maxLength: 1000 }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: { type: string }
                  status: { type: string, enum: [updated] }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [saved]
      operationId: unsaveCompany
      summary: Remove a company from the caller's watchlist.
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: company_id
          required: true
          schema:
            type: string
            pattern: "^co_[a-z0-9]+$"
      responses:
        "204": { description: Unsaved. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  responses:
    BadRequest:
      description: Validation error or malformed request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: Authenticated but not authorized for this resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth.session_token
      description: |
        Better Auth session cookie. Set by `/sign-in` or `/sign-up`.
        Roles: `owner`, `goeo_admin`, `superadmin`. Owner edits gate
        on `companies.claimed_by_user_id` matching `user.id`.
    atlasAdminToken:
      type: apiKey
      in: header
      name: X-Atlas-Admin-Token
      description: |
        Service-account-style token validated against
        `env.ATLAS_ADMIN_TOKEN`. The local CLI (`npm run cli` from a
        checkout), the local stdio MCP server (`npm run mcp`), and
        external machine clients use this header for write operations.
        Bypasses Better Auth and assumes a privileged caller.
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: bad_request }
            message: { type: string }
            details: {}
...
    InvestorPublicCard:
      type: object
      required:
        - id
        - slug
        - display_name
        - stages
        - sectors
        - geo_focus
        - verification
        - canonical_url
        - agent_card_urls
      properties:
        id: { type: string, pattern: "^inv_[a-z0-9]+$" }
        slug: { type: string }
        display_name: { type: string }
        firm_name: { type: string, nullable: true }
        investor_type: { type: string, nullable: true }
        investor_type_display: { type: string, nullable: true }
        tagline: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        website: { type: string, format: uri, nullable: true }
        linkedin: { type: string, format: uri, nullable: true }
        stages:
          type: array
          items: { type: string }
        stages_display:
          type: array
          items: { type: string }
        sectors:
          type: array
          items: { type: string }
        sectors_display:
          type: array
          items: { type: string }
        check_size_min: { type: integer, nullable: true }
        check_size_max: { type: integer, nullable: true }
        geo_focus:
          type: array
          items: { type: string }
        geo_focus_display:
          type: array
          items: { type: string }
        verification:
          type: object
          required: [status]
          properties:
            status: { type: string, enum: [verified, unverified] }
            verified_at: { type: string, format: date-time, nullable: true }
        last_updated_at: { type: string, format: date-time, nullable: true }
        canonical_url: { type: string }
        agent_card_urls:
          type: object
          required: [html, markdown, json, api]
          properties:
            html: { type: string }
            markdown: { type: string }
            json: { type: string }
            api: { type: string }
    InvestorPublicPatch:
      type: object
      description: |
        All fields optional. Owner edits are limited to the public
        whitelist (display_name, bio, tagline, website, linkedin, slug
        when unverified). `verification_status` is admin/machine only.
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"
          maxLength: 80
        display_name: { type: string, maxLength: 120 }
        bio: { type: string, nullable: true, maxLength: 2000 }
        tagline: { type: string, nullable: true, maxLength: 200 }
        website: { type: string, format: uri, nullable: true }
        linkedin: { type: string, format: uri, nullable: true }
        verification_status:
          type: string
          enum: [unverified, verified]
          description: Admin / machine only.
    SavedCompany:
      type: object
      required: [id, saved_at]
      properties:
        id: { type: string, pattern: "^sc_[a-z0-9]+$" }
        saved_at: { type: string, format: date-time, nullable: true }
        note: { type: string, nullable: true }
        company:
          type: object
          nullable: true
          properties:
            id: { type: string }
            slug: { type: string }
            name: { type: string }
            sector: { type: string, nullable: true }
            stage: { type: string, nullable: true }
            logo_url: { type: string, nullable: true }
...